review fixes: - remove python version in README - refactor user serializers, model @
