+++ README.md
@@ -31,7 +31,7 @@ git clone git@github.com:bpaverin/api_final_yatube.git
 Cоздать и активировать виртуальное окружение:
 
 ```
-python3 -m venv env
+python -m venv env
 ```
 
 ```
@@ -41,7 +41,7 @@ source env/bin/activate
 Установить зависимости из файла requirements.txt:
 
 ```
-python3 -m pip install --upgrade pip
+python -m pip install --upgrade pip
 ```
 
 ```
@@ -51,19 +51,19 @@ pip install -r requirements.txt
 Выполнить миграции:
 
 ```
-python3 manage.py migrate
+python manage.py migrate
 ```
 
 Предусмотрена возможность загрузить заранее подготовленные данные для таблиц:
 ```
 cd static
-python3.9 data_loader.py 
+python data_loader.py 
 ```
 
 Запустить проект:
 
 ```
-python3 manage.py runserver
+python manage.py runserver
 ```
 
 ### Примеры запросов
