+++ api_yamdb/api/permissions.py
@@ -3,11 +3,7 @@ from rest_framework import permissions
 
 class IsAdminOrSuperUser(permissions.BasePermission):
     def has_permission(self, request, view):
-        return request.user.is_authenticated and (
-            request.user.is_admin
-            or request.user.is_superuser
-            or request.user.is_staff
-        )
+        return request.user.is_authenticated and (request.user.is_admin)
 
 
 class IsAdminModerOrAuthor(permissions.IsAuthenticatedOrReadOnly):
@@ -16,8 +12,7 @@ class IsAdminModerOrAuthor(permissions.IsAuthenticatedOrReadOnly):
             request.method in permissions.SAFE_METHODS
             or request.user.is_authenticated
             and (
-                request.user.is_superuser
-                or request.user.is_admin
+                request.user.is_admin
                 or request.user.is_moderator
                 or request.user == obj.author
             )
