+++ api_yamdb/api/serializers.py
@@ -4,25 +4,21 @@ from django.core.mail import send_mail
 from django.shortcuts import get_object_or_404
 from rest_framework import serializers
 from rest_framework_simplejwt.serializers import TokenObtainSerializer
-from rest_framework_simplejwt.tokens import RefreshToken
+from rest_framework_simplejwt.tokens import AccessToken
 
 from reviews.models import Category, Comment, Genre, Review, Title, YamdbUser
-from yamdb_user.models import (
-    EMAIL_MAX_LENGTH,
-    USERNAME_MAX_LENGTH,
-    BaseUserValidator,
-)
+from yamdb_user.models import EMAIL_MAX_LENGTH, USERNAME_MAX_LENGTH
+from yamdb_user.validators import BaseUserValidator
 
 
 class CustomTokenObtainPairSerializer(TokenObtainSerializer):
+    token_class = AccessToken
+
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         del self.fields['password']
         self.fields['username'] = serializers.CharField()
         self.fields['confirmation_code'] = serializers.CharField()
-
-    def get_token(cls, user):
-        return RefreshToken.for_user(user)
 
     def validate(self, attrs):
         username = attrs['username']
@@ -44,11 +40,12 @@ class UserSerializerAuth(serializers.Serializer, BaseUserValidator):
         username = attrs.get('username')
         email = attrs.get('email')
         errors = {}
-        email_exists = YamdbUser.objects.filter(email=email).exists()
-        username_exists = YamdbUser.objects.filter(username=username).exists()
-        if (
-            not email_exists and not username_exists
-        ) or YamdbUser.objects.filter(email=email, username=username).exists():
+        email_exists = YamdbUser.objects.filter(email=email).first()
+        username_exists = YamdbUser.objects.filter(username=username).first()
+        # if (
+        #     not email_exists and not username_exists
+        # ) or YamdbUser.objects.filter(email=email, username=username).exists():
+        if email_exists == username_exists:
             return attrs
         if email_exists:
             errors['email'] = 'Email already exists'
